fix: Make deployment test more robust
We do not know which keys are set from outside so ignore additional keys.
Also make sure that if a value is None it is the same as if the key is missing,
helm templates has some special handling here if a key is overriden with null. If
it was available previously the key will be removed but if it was not available the
key will be available with value None.

univention/dev/internal/team-nubus#1096