refactor(helm-test-harness): Move "client_udm" into "helm-test-harness"
The package has been created when working on the portal before we did have a
solution in place to package up the "helm-test-harness" and use it in other
projects.

This refactoring moves the library into the test harness, so that the duplicated
code in the portal repository can be removed entirely.